Criando services e conversores

+++ src/main/java/com/rhpro/dto/inputs/LicencaInput.java
@@ -19,5 +19,5 @@ public class LicencaInput {
     //é pra ser uma imagem...
     private File atestado;
 
-    private Integer funcionarioId;
+    private Long funcionarioId;
 }
